tests/wpt/web-platform-tests/html/semantics/scripting-1/the-script-element/execution-timing/015.html

<!DOCTYPE html>
<html><head>
        <title> scheduler: DOM added inline+external+inline script earlier in document</title>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="testlib/testlib.js"></script>
</head>
<body>

        <div id="log">FAILED (This TC requires JavaScript enabled)</div>

        <script>log('inline script #1');
                testlib.addScript('log(\'head script #1\')', {}, document.getElementsByTagName('head')[0], false);
                var s = testlib.addScript('', { 'src':'scripts/include-1.js' }, document.getElementsByTagName('head')[0], false);
                testlib.addScript('log(\'head script #2\')', {}, document.getElementsByTagName('head')[0], false);
                log('end script #1');
        </script>

        <script type="text/javascript">
        log( 'inline script #2' );
        var t = async_test()

        ///XXX I think the spec allows this case to race
        onload = function(){
          setTimeout(t.step_func(
            function() {
              assert_any(assert_array_equals, eventOrder, [['inline script #1', 'head script #1', 'head script #2', 'end script #1', 'external script #1', 'inline script #2'],
                                                           ['inline script #1', 'head script #1', 'head script #2', 'end script #1', 'inline script #2', 'external script #1']]);
              t.done();
            }),
          100);}
        </script>

</body></html>
